refactor: consolidate files
